Moved InlineTextInput to BotCommands-jda-ktx
As it is now no longer reliant on BC's modals
